some comments fixed
