+++ tests/conftest.py
@@ -1,15 +1,18 @@
 import pytest
 from selenium import webdriver
 from selenium.webdriver.common.by import By
-from selenium.webdriver.common.keys import Keys
 
-def test_admin_login(base_url, selenium):
+@pytest.fixture
+def login(base_url, selenium):
     selenium.get('{0}/admin/login/'.format(base_url))
     selenium.maximize_window()
+
+    # login as admin
     username = selenium.find_element_by_id('id_username')
     password = selenium.find_element_by_id('id_password')
-    submit = selenium.find_element_by_class_name('submit-row')
+    login = selenium.find_element_by_class_name('submit-row')
     username.send_keys("admin")
     password.send_keys("admin")
-    submit.click()
-    assert selenium.title == "Site administration | Django site admin"
+    login.click()
+
+    return login
